move constants to constants
